pyproject.toml
[project]
name = "feedee"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "djangorestframework>=3.15.2",
    "feedparser>=6.0.11",
    "requests>=2.32.3",
    "ruff>=0.6.4",
]
